refactor: separate quiz-session into another file
