+++ Main/src/page/quiz/solve.ts
@@ -1,5 +1,5 @@
 import "../../../styles/quiz";
-import { QuizSession } from "../../api/quiz";
+import { QuizSession } from "../../api/quiz_session";
 import {
     addAnswerSubmitListener,
     displayCorrectnessAnimation,
